Move 'Header' component to 'shared' for better structure
