--- src/components/layout/Layout.js
@@ -1,13 +1,0 @@
-import React from 'react';
-// import Header from './Header';
-// import Navbar from '../navigation/Navbar';
-// import Footer from '../footer/Footer';
-// import { Outlet } from 'react-router-dom';
-
-// Leave this for now, I may use it later
-
-const Layout = () => {
-  return <div></div>;
-};
-
-export default Layout;
