+++ src/views/User/profile/ProfileForm.js
@@ -13,7 +13,7 @@ import {
   RadioGroup,
   Radio,
 } from '@mui/material';
-import Header from '../../../components/layout/Header';
+import Header from '../../../components/shared/Header';
 
 const initialState = {
   username: '',
