+++ src/views/error/ErrorPage.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Header from '../../components/layout/Header';
+import Header from '../../components/shared/Header';
 
 function ErrorPage() {
   return (
