+++ src/views/post/NewPost.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Header from '../../components/layout/Header';
+import Header from '../../components/shared/Header';
 import PostForm from '../../components/posts/forms/PostForm';
 
 // MUI
